Extract action_view instrumentations
