+++ lib/ddtrace/contrib/action_view/patcher.rb
@@ -1,7 +1,8 @@
 require 'ddtrace/ext/http'
 require 'ddtrace/contrib/patcher'
 require 'ddtrace/contrib/action_view/ext'
-require 'ddtrace/contrib/action_view/instrumentation'
+require 'ddtrace/contrib/action_view/instrumentations/partial_renderer'
+require 'ddtrace/contrib/action_view/instrumentations/template_renderer'
 require 'ddtrace/contrib/action_view/utils'
 
 module Datadog
@@ -31,16 +32,16 @@ module Datadog
           do_once(:patch_renderer) do
             if defined?(::ActionView::TemplateRenderer) && defined?(::ActionView::PartialRenderer)
               if Integration.version < Gem::Version.new('6.0.0')
-                ::ActionView::TemplateRenderer.send(:prepend, Instrumentation::TemplateRenderer::Rails31To5)
-                ::ActionView::PartialRenderer.send(:prepend, Instrumentation::PartialRenderer::RailsLessThan6)
+                ::ActionView::TemplateRenderer.send(:prepend, Instrumentations::TemplateRenderer::Rails31To5)
+                ::ActionView::PartialRenderer.send(:prepend, Instrumentations::PartialRenderer::RailsLessThan6)
               else
-                ::ActionView::TemplateRenderer.send(:prepend, Instrumentation::TemplateRenderer::Rails6Plus)
-                ::ActionView::PartialRenderer.send(:prepend, Instrumentation::PartialRenderer::Rails6Plus)
+                ::ActionView::TemplateRenderer.send(:prepend, Instrumentations::TemplateRenderer::Rails6Plus)
+                ::ActionView::PartialRenderer.send(:prepend, Instrumentations::PartialRenderer::Rails6Plus)
               end
             elsif defined?(::ActionView::Rendering) && defined?(::ActionView::Partials::PartialRenderer)
               # NOTE: Rails < 3.1 compatibility: different classes are used
-              ::ActionView::Rendering.send(:prepend, Instrumentation::TemplateRenderer::Rails30)
-              ::ActionView::Partials::PartialRenderer.send(:prepend, Instrumentation::PartialRenderer::RailsLessThan6)
+              ::ActionView::Rendering.send(:prepend, Instrumentations::TemplateRenderer::Rails30)
+              ::ActionView::Partials::PartialRenderer.send(:prepend, Instrumentations::PartialRenderer::RailsLessThan6)
             else
               Datadog::Tracer.log.debug('Expected Template/Partial classes not found; template rendering disabled')
             end
