+++ spec/ddtrace/contrib/active_record/multi_db_spec.rb
@@ -98,7 +98,7 @@ RSpec.describe 'ActiveRecord multi-database implementation' do
 
     context 'a Symbol that matches a configuration' do
       context 'when ActiveRecord has configurations' do
-        let(:configurations) do
+        let(:database_configuration) do
           {
             'gadget' => {
               'encoding' => 'utf8',
@@ -118,16 +118,18 @@ RSpec.describe 'ActiveRecord multi-database implementation' do
           }
         end
 
-        before(:each) do
-          # Stub ActiveRecord::Base, to pretend its been configured
+        let(:database_configuration_object) do
           if defined?(ActiveRecord::DatabaseConfigurations)
-            allow(ActiveRecord::Base).to receive(:configurations).and_return(
-              ActiveRecord::DatabaseConfigurations.new(configurations)
-            )
+            ActiveRecord::DatabaseConfigurations.new(database_configuration)
           else
             # Before Rails 6, database configuration was a plain Hash
-            allow(ActiveRecord::Base).to receive(:configurations).and_return(configurations)
+            database_configuration
           end
+        end
+
+        before(:each) do
+          # Stub ActiveRecord::Base, to pretend its been configured
+          allow(ActiveRecord::Base).to receive(:configurations).and_return(database_configuration_object)
 
           Datadog.configure do |c|
             c.use :active_record, describes: :gadget do |gadget_db|
